requirements.txt
python-telegram-bot==20.7
gspread==6.1.2
oauth2client==4.1.3
aiohttp==3.9.5
google-auth==2.35.0
nest-asyncio==1.6.3